TKSS-526: Remove SM2Operations
