+++ kona-crypto/src/main/java/com/tencent/kona/crypto/provider/SM2KeyPairGenerator.java
@@ -37,8 +37,8 @@ import java.security.spec.AlgorithmParameterSpec;
 import java.security.spec.ECPoint;
 import java.util.Arrays;
 
-import static com.tencent.kona.sun.security.ec.SM2Operations.SM2OPS;
-import static com.tencent.kona.sun.security.ec.SM2Operations.toECPoint;
+import static com.tencent.kona.sun.security.ec.ECOperations.SM2OPS;
+import static com.tencent.kona.sun.security.ec.ECOperations.toECPoint;
 
 public final class SM2KeyPairGenerator extends KeyPairGeneratorSpi {
