--- kona-crypto/src/main/java/com/tencent/kona/sun/security/ec/SM2Operations.java
@@ -1,77 +1,0 @@
-/*
- * Copyright (C) 2023, THL A29 Limited, a Tencent company. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2 as
- * published by the Free Software Foundation. THL A29 Limited designates
- * this particular file as subject to the "Classpath" exception as provided
- * in the LICENSE file that accompanied this code.
- *
- * This code is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License version 2 for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-
-package com.tencent.kona.sun.security.ec;
-
-import com.tencent.kona.crypto.CryptoUtils;
-import com.tencent.kona.crypto.spec.SM2ParameterSpec;
-import com.tencent.kona.sun.security.ec.point.AffinePoint;
-import com.tencent.kona.sun.security.ec.point.Point;
-import com.tencent.kona.sun.security.util.math.IntegerFieldModuloP;
-import com.tencent.kona.sun.security.util.math.IntegerModuloP;
-import com.tencent.kona.sun.security.util.math.intpoly.IntegerPolynomialSM2;
-import com.tencent.kona.sun.security.util.math.intpoly.SM2OrderField;
-
-import java.security.MessageDigest;
-import java.security.SecureRandom;
-import java.security.spec.ECPoint;
-
-/**
- * Elliptic curve point arithmetic for SM2.
- */
-public class SM2Operations extends ECOperations {
-
-    // limit = order - 1
-    private static final byte[] LIMIT = CryptoUtils.toBytesLE(
-            "FFFFFFFEFFFFFFFFFFFFFFFFFFFFFFFF7203DF6B21C6052B53BBF40939D54122");
-
-    public static final SM2Operations SM2OPS = new SM2Operations(
-            IntegerPolynomialSM2.ONE.getElement(SM2ParameterSpec.CURVE.getB()),
-            SM2OrderField.ONE);
-
-    public SM2Operations(IntegerModuloP b, IntegerFieldModuloP orderField) {
-        super(b, orderField);
-    }
-
-    @Override
-    public byte[] generatePrivateScalar(SecureRandom random) {
-        byte[] privArr = super.generatePrivateScalar(random);
-
-        // If the privArr is order - 1, just try once again.
-        if (MessageDigest.isEqual(LIMIT, privArr)) {
-            // It should unlikely get here
-            privArr = super.generatePrivateScalar(random);
-        }
-
-        return privArr;
-    }
-
-    public static ECPoint toECPoint(Point point) {
-        AffinePoint affPoint = point.asAffine();
-        return new ECPoint(
-                affPoint.getX().asBigInteger(),
-                affPoint.getY().asBigInteger());
-    }
-
-    public static boolean isInfinitePoint(Point point) {
-        AffinePoint affPoint = point.asAffine();
-        return affPoint.getX() == null || affPoint.getY() == null;
-    }
-}
